+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [ next ]
+### Fixed
+- empkg: acl: fix recursive path handling
+
 ## [ 5.1.0 ] - 2024-08-19
 ### Changed
 - empkg: ported to application written in C